Building a DAO layer

+++ Chapter04/MySQL/Maven/SimpleDao/src/main/java/com/classicmodels/repository/SaleRepositoryImpl.java
@@ -24,10 +24,10 @@ public class SaleRepositoryImpl implements SaleRepository {
     }
 
     @Override
-    public List<Sale> findSaleAscGtLimit(int limit) {
+    public List<Sale> findSaleAscGtLimit(double limit) {
 
         return ctx.selectFrom(SALE)
-                .where(SALE.SALE_.ge(Double.valueOf(limit)))
+                .where(SALE.SALE_.ge(limit))
                 .orderBy(SALE.SALE_)
                 .fetchInto(Sale.class);
     }
